Renamed Classes
Renamed classes for a better overview
