+++ JoinMusik/src/plugin.yml
@@ -1,10 +1,12 @@
 name: JoinMusik
-author: T0biii
+authors:
+  - T0biii
+  - Fruxz
 depend: [NoteBlockAPI]
-version: 2.2
+version: '2.2'
 website: http://dev.bukkit.org/bukkit-plugins/joinmusik
 
-main: de.t0biii.musik.JoinMusic
+main: de.t0biii.musik.main.Main
 database: false
 commands:
   JoinMusik:
